+++ src/test/java/spitter/WriterRunnerWithNiceWriterTest.java
@@ -3,26 +3,24 @@ package spitter;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.omg.CORBA.IRObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import spitter.web.services.writer.IWriter;
 import spitter.web.services.writer.NiceWriter;
-import spitter.web.services.writer.Writer;
 import spitter.web.services.writer.WriterRunner;
 
 /**
  * Created by nkim on 5/9/2017.
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = {WriterTestConfig.class})
-public class UnitTestWriterRunnerWithWriter {
+@ContextConfiguration(classes = {NiceWriterTestConfig.class})
+public class WriterRunnerWithNiceWriterTest {
     @Autowired
     private WriterRunner writerRunner;
 
     @Autowired
-    private IWriter iwriter;
+    private IWriter writer;
 
     @Test
     public void getWriterRunnerObject_ReturnObject(){
@@ -30,23 +28,23 @@ public class UnitTestWriterRunnerWithWriter {
     }
 
     @Test
-    public void IWriterNotNull(){
-        Assert.assertNotNull(iwriter);
+    public void checkIWriter_returnObject(){
+        Assert.assertNotNull(writer);
     }
 
     @Test
-    public void checkIWriter_isInstanceWriter(){
-        Assert.assertTrue(iwriter instanceof Writer);
+    public void checkIWriter_returnInstanceNiceWriter(){
+        Assert.assertTrue(writer instanceof NiceWriter);
     }
 
     @Test
-    public void GetWriterMethod_returnIWriter(){
-        Assert.assertEquals(writerRunner.getWriter(), iwriter);
+    public void getWriterMethod_returnIWriter(){
+        Assert.assertEquals(writerRunner.getWriter(), writer);
     }
 
     @Test
-    public void getWriter_ReturnWriterObject(){
-        Assert.assertTrue(writerRunner.getWriter() instanceof Writer);
+    public void getWriter_ReturnNiceWriterObject(){
+        Assert.assertTrue(writerRunner.getWriter() instanceof NiceWriter);
     }
 
 }
