+++ src/test/java/spitter/WriterRunnerWithWriterTest.java
@@ -3,6 +3,7 @@ package spitter;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.omg.CORBA.IRObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -15,8 +16,8 @@ import spitter.web.services.writer.WriterRunner;
  * Created by nkim on 5/9/2017.
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = {NiceWriterTestConfig.class})
-public class UnitTestWriterRunnerWithNiceWriter {
+@ContextConfiguration(classes = {WriterTestConfig.class})
+public class WriterRunnerWithWriterTest {
     @Autowired
     private WriterRunner writerRunner;
 
@@ -29,23 +30,25 @@ public class UnitTestWriterRunnerWithNiceWriter {
     }
 
     @Test
-    public void IWriterNotNull(){
+    public void checkIWriter_returnObject(){
         Assert.assertNotNull(iwriter);
     }
 
     @Test
-    public void checkIWriter_isInstanceNiceWriter(){
-        Assert.assertTrue(iwriter instanceof NiceWriter);
+    public void checkIWriter_returnInstanceWriter(){
+        Assert.assertTrue(iwriter instanceof Writer);
     }
 
     @Test
-    public void GetWriterMethod_returnIWriter(){
+    public void getWriterMethod_returnIWriter(){
         Assert.assertEquals(writerRunner.getWriter(), iwriter);
     }
 
     @Test
-    public void getWriter_ReturnNiceWriterObject(){
-        Assert.assertTrue(writerRunner.getWriter() instanceof NiceWriter);
+    public void getWriterMethod_ReturnWriterObject(){
+        Assert.assertTrue(writerRunner.getWriter() instanceof Writer);
     }
+
+
 
 }
